+++ src/test/java/rest/persistence/repository/ObjectRepositoryTest.java
@@ -20,6 +20,7 @@ import rest.BaseWebTest;
 
 import java.io.FileNotFoundException;
 import java.io.IOException;
+import java.math.BigInteger;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
@@ -144,13 +145,13 @@ public class ObjectRepositoryTest extends BaseWebTest {
 
         ReferenceEntry entry = anno4j.createObject(ReferenceEntry.class);
         this.entryId1 = entry.getResourceAsString();
-        entry.setPages(11);
+        entry.setPages(BigInteger.valueOf(11));
 
         entry.setIsAbout(group);
         group.addEntry(entry);
 
         ReferenceEntry entry2 = anno4j.createObject(ReferenceEntry.class);
-        entry2.setPages(22);
+        entry2.setPages(BigInteger.valueOf(22));
 
         entry2.setIsAbout(group);
         group.addEntry(entry2);
